Refactored lists in dictionarys
