+++ main.py
@@ -12,7 +12,7 @@ from src.common_constants import CommonConstants, GameFonts, Opacity
 from src.overlay import Overlay
 from src.game_timing import GameTiming
 from src.landing_game_action_on_key import PygameKeyState, LandingGameActionOnKey
-from src.object_identifier import ObjectIdentifier
+from src.object_library import ObjectIdentifier
 
 
 def create_pg_surface_from_color_and_size(color, size):
@@ -175,11 +175,6 @@ def main():
         PygameKeyState(pygame.K_v, True), toggle_overlay_visibility_on_down
     )
 
-    obj_list = pygame.sprite.Group()
-    obj_list.add(key_indicator_while_pressed)
-    obj_list.add(key_indicator_on_down)
-    obj_list.add(ego)
-    obj_list.add(ground)
     for overlay in overlays.values():
         obj_list.add(overlay)
